requirements.txt
python-fslogger>=0.1.4
python-fssignal>=0.0.7
python-fspacker>=0.1.0